StructuralUnits

Documentation for StructuralUnits.

StructuralUnits.to_deg

StructuralUnits.to_ft

StructuralUnits.to_inch

StructuralUnits.to_inch2

StructuralUnits.to_kcf

StructuralUnits.to_klf

StructuralUnits.to_ksi

StructuralUnits.to_mph

StructuralUnits.to_pcf

StructuralUnits.to_plf

StructuralUnits.to_deg — Method

to_deg(x)

Converts x to a Unitful quantity with unit deg.

StructuralUnits.to_ft — Method

to_ft(x)

Converts x to a Unitful quantity with unit ft.

StructuralUnits.to_inch — Method

to_inch(x)

Converts x to a Unitful quantity with unit inch.

StructuralUnits.to_inch2 — Method

to_inch2(x)

Converts x to a Unitful quantity with unit inch^2.

StructuralUnits.to_kcf — Method

to_kcf(x)

Converts x to a Unitful quantity with unit kcf.

StructuralUnits.to_klf — Method

to_klf(x)

Converts x to a Unitful quantity with unit plf.

StructuralUnits.to_ksi — Method

to_ksi(x)

Converts x to a Unitful quantity with unit ksi.

StructuralUnits.to_mph — Method

to_mph(x)

Converts x to a Unitful quantity with unit mph.

StructuralUnits.to_pcf — Method

to_pcf(x)

Converts x to a Unitful quantity with unit pcf.

StructuralUnits.to_plf — Method

to_plf(x)

Converts x to a Unitful quantity with unit plf.